requirements.txt
# Inside requirements.txt
google-generativeai
fastapi
streamlit